+++ src/main/java/org/doancnpm/DAO/PhieuNhapDAO.java
@@ -100,7 +100,7 @@ public class PhieuNhapDAO implements Idao<PhieuNhap> {
             String nhaCungCap = rs.getString("nhaCungCap");
             int maNhanVien = rs.getInt("MaNhanVien");
             Date ngayLap  =rs.getDate("NgayLapPhieu");
-            double tongTien = rs.getDouble("tongTien");
+            Long tongTien = rs.getLong("tongTien");
             String ghiChu = rs.getString("GhiChu");
 
             phieuNhap = new PhieuNhap();
@@ -133,7 +133,7 @@ public class PhieuNhapDAO implements Idao<PhieuNhap> {
             String nhaCungCap = rs.getString("nhaCungCap");
             int maNhanVien = rs.getInt("MaNhanVien");
             Date ngayLap  =rs.getDate("NgayLapPhieu");
-            double tongTien = rs.getDouble("tongTien");
+            Long tongTien = rs.getLong("tongTien");
             String ghiChu = rs.getString("GhiChu");
 
             phieuNhap = new PhieuNhap();
@@ -165,7 +165,7 @@ public class PhieuNhapDAO implements Idao<PhieuNhap> {
             String nhaCungCap = rs.getString("nhaCungCap");
             int maNhanVien = rs.getInt("MaNhanVien");
             Date ngayLap  =rs.getDate("NgayLapPhieu");
-            double tongTien = rs.getDouble("tongTien");
+            Long tongTien = rs.getLong("tongTien");
             String ghiChu = rs.getString("GhiChu");
 
             PhieuNhap phieuNhap = new PhieuNhap();
