+++ src/main/java/org/doancnpm/Models/ChiTietPhieuNhap.java
@@ -4,7 +4,7 @@ public class ChiTietPhieuNhap {
     Integer maPhieuNhap;
     Integer maMatHang;
     Integer soLuong;
-    Double thanhTien;
+    Long thanhTien;
 
     public Integer getMaPhieuNhap() {
         return maPhieuNhap;
@@ -30,11 +30,11 @@ public class ChiTietPhieuNhap {
         this.soLuong = soLuong;
     }
 
-    public Double getThanhTien() {
+    public Long getThanhTien() {
         return thanhTien;
     }
 
-    public void setThanhTien(Double thanhTien) {
+    public void setThanhTien(Long thanhTien) {
         this.thanhTien = thanhTien;
     }
